requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
click==8.1.7
distro==1.9.0
fastapi==0.115.6
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
loguru==0.7.3
openai==1.57.4
packaging==24.2
pluggy==1.5.0
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.3.4
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.19
sniffio==1.3.1
starlette==0.41.3
tenacity==9.0.0
tqdm==4.67.1
typeguard==4.4.1
typing_extensions==4.12.2
uvicorn==0.34.0
